chore(app): change tx utils
