--- packages/app/src/systems/Transaction/utils/parse.tsx
@@ -1,9 +1,0 @@
-import type { TxRequest, TxResponse } from '../types';
-
-export function parseTransaction<T extends TxRequest | TxResponse>(tx: T) {
-  return Object.entries(tx).reduce((obj, [key, value]) => {
-    const val =
-      value instanceof Uint8Array ? value : JSON.parse(JSON.stringify(value));
-    return { ...obj, [key]: val };
-  }, {} as T);
-}
